[FIX]pref: do not register in databse pref change - fix calendar visible for anonymous
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@16087 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/tikilib.php
@@ -4606,12 +4606,14 @@ class TikiLib extends TikiDB {
 			}
 		}
 
-		$query = "delete from `tiki_user_preferences` where `user`=? and `prefName`=?";
-		$bindvars=array($my_user,$name);
-		$result = $this->query($query, $bindvars, -1,-1,false);
-		$query = "insert into `tiki_user_preferences`(`user`,`prefName`,`value`) values(?, ?, ?)";
-		$bindvars[]=$value;
-		$result = $this->query($query, $bindvars);
+		if (!empty($my_user)) {
+			$query = "delete from `tiki_user_preferences` where `user`=? and `prefName`=?";
+			$bindvars=array($my_user,$name);
+			$result = $this->query($query, $bindvars, -1,-1,false);
+			$query = "insert into `tiki_user_preferences`(`user`,`prefName`,`value`) values(?, ?, ?)";
+			$bindvars[]=$value;
+			$result = $this->query($query, $bindvars);
+		}
 
 		return true;
 	}
